Mainly temporary modifications, not stable. Email template and sending solution added, controllers updated continously

+++ src/controllers/review.js
@@ -1,32 +1,33 @@
-import Product from "../models/product.js";
 import Review from "../models/review.js";
-import { isValidObjectId } from "../utils/mongoIdValidation.js";
+import { hasPaginationParams } from "../utils/controllerUtils/general.js";
+import {
+  hasUpdateProps,
+  reviewExists,
+} from "../utils/controllerUtils/review.js";
 
 export const getReviewsByUserId = async (req, res) => {
-  const { query, userId } = req;
-  const page = parseInt(query.page);
-  const limit = parseInt(query.limit);
-  const pageCondition = page && Number.isInteger(page) && page > 0;
-  const limitCondition =
-    limit && Number.isInteger(limit) && limit > 0 && limit <= 100;
+  const {
+    query: { page, limit },
+    userId,
+  } = req;
 
   try {
-    if (pageCondition && limitCondition) {
+    if (hasPaginationParams(page, limit)) {
       const reviews = await Review.find({ userId })
-        .select(["-createdAt", "-updatedAt", "-__v", "-userId"])
+        .select(["-__v", "-userId"])
         .populate({
           path: "productId",
           model: "Product",
           select: ["name", "thumbnail"],
         })
-        .sort({ _id: 1 })
+        .sort({ createdAt: 1 })
         .skip((page - 1) * limit)
-        .limit(10);
+        .limit(limit);
       res.status(200).json(reviews);
     } else {
-      res.status(500).json({
-        message: "The pageNumber must be an integer greater than zero",
-      });
+      throw new Error(
+        "The page and limit parameters must be integers greater than zero"
+      );
     }
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -35,33 +36,27 @@ export const getReviewsByUserId = async (req, res) => {
 
 export const getReviewsByProductId = async (req, res) => {
   const {
-    query,
+    query: { page, limit },
     body: { productId },
   } = req;
-  const page = parseInt(query.page);
-  const limit = parseInt(query.limit);
-  const pageCondition = page && Number.isInteger(page) && page > 0;
-  const limitCondition =
-    limit && Number.isInteger(limit) && limit > 0 && limit <= 100;
 
   try {
-    if (pageCondition && limitCondition) {
+    if (hasPaginationParams(page, limit)) {
       const reviews = await Review.find({ productId })
-        .select(["text", "rating", "userId"])
+        .select(["-__v", "-productId"])
         .populate({
           path: "userId",
           model: "User",
           select: ["fullNameHUN", "fullNameENG", "image"],
         })
-        .sort({ _id: 1 })
+        .sort({ createdAt: 1 })
         .skip((page - 1) * limit)
         .limit(limit);
       res.status(200).json(reviews);
     } else {
-      res.status(500).json({
-        message:
-          "The page and limit parameters must be integer convertable strings greater than zero",
-      });
+      throw new Error(
+        "The page and limit parameters must be integers greater than zero"
+      );
     }
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -75,18 +70,12 @@ export const createReview = async (req, res) => {
   } = req;
 
   try {
-    if (isValidObjectId(productId)) {
-      const isValidProductId = await Product.exists({ _id: productId });
-      if (isValidProductId) {
-        await Review.create({ text, rating, userId, productId });
-        res.status(200).json({ message: "Review created" });
-      } else {
-        res.status(500).json({ message: "Provided productId isn't valid" });
-      }
+    if (await productExists(productId)) {
+      const product = { text, rating, userId, productId };
+      await Review.create(product);
+      res.status(200).json({ message: "Review created" });
     } else {
-      res
-        .status(500)
-        .json({ message: "Provided productId isn't a valid ObjectId" });
+      throw new Error("There isn't a product with this id");
     }
   } catch (err) {
     res.status(500).json({ message: err.message });
@@ -99,8 +88,12 @@ export const deleteReviewById = async (req, res) => {
   } = req;
 
   try {
-    await Review.findByIdAndDelete(id);
-    res.status(200).json({ message: "Review deleted" });
+    if (await reviewExists(id)) {
+      await Review.findByIdAndDelete(id);
+      res.status(200).json({ message: "Review deleted" });
+    } else {
+      throw new Error("There isn't a review with this id");
+    }
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
@@ -111,20 +104,18 @@ export const updateReviewById = async (req, res) => {
     params: { id },
     body: { text, rating },
   } = req;
-  const ratingCondition =
-    rating && Number.isInteger(rating) && !(rating < 1) && !(rating > 5);
 
   try {
-    const review = { text };
-    if (ratingCondition) review.rating = rating;
-    if (text || text === "") {
-      await Review.findByIdAndUpdate(id, review);
-      res.status(200).json({ message: "Review updated" });
+    if (await reviewExists(id)) {
+      const review = { text, rating };
+      if (hasUpdateProps(review)) {
+        await Review.findByIdAndUpdate(id, review, { runValidators: true });
+        res.status(200).json({ message: "Review updated" });
+      } else {
+        throw new Error("Rating and text are required for review update");
+      }
     } else {
-      res.status(500).json({
-        message:
-          "No data provided for review update or the data isn't acceptable",
-      });
+      throw new Error("There isn't a review with this id");
     }
   } catch (err) {
     res.status(500).json({ message: err.message });
